arc.Fsp: Close CrOS Files App after unzipping
In arc.Fsp, Files App is launched twice with filesapp.Launch() to first
unzip a test zip file, and to second open a file with an Android app.
With the switching of default Files App used in tast test from ChromeApp
to SWA in crrev/c/3849961, this now results in opening two windows for
Files App, causing the UI Automator to fail to find the window.
This CL fixes the test by closing the first Files App window after the
unzip.

BUG=b:244301416
TEST=tast run <DUT> arc.Fsp.*

Change-Id: I69e415bf67546b84f5e0ec85fae141a106065b9f