fusebox: remove FuseBoxReverseService-related tests
The original FuseBox design used two D-Bus interfaces:
- FBS  (org.chromium.FuseBoxService)
- FBRS (org.chromium.FuseBoxReverseService)

For FBS, the Chrome process is the D-Bus server and the FuseBox process
is the D-Bus client. This is swapped for FBRS, so FuseBox is the server
and Chrome is the client.

We would like to experiment with simplifying or even removing the FBRS
interface, which would allow the FBS server (i.e. Chrome) to serve
multiple clients simultaneously, not just one.

This commit removes test code that previously relied on the FBRS
interface. It is not easily replaced, as "call the FBRS.AttachStorage
method" has been deprecated in favor of "listen to the
FBS.StorageAttached signal" but while the fusebox_attach_storage.go test
(prior to this commit) could call into *the* FBRS server by the
equivalent of shelling out to dbus-send, it is harder to fake an FBS
signal (from outside *the* FBS process).

BUG=chromium:1249754
TEST=tast run DUT platform.Fusebox*

Change-Id: I04affafba20b8888055e47b2031e96a92853e9c4