crostini: stabilize remove.* on low-end devices
The remove button seems to be unstable in the a11y tree when first
rendered in the tree, e.g., in https://screenshot.googleplex.com/4quD5gAmqPT3tY8
it finds the "port forwarding" as the location for the remove button.

Add a 1s sleep should fix the problem.
Got consistent failures on drawlat device without sleep, and consistent
passes with the sleep.

BUG=b:241565887
TEST=tast run <DUT> crostini.RemoveCancel*

Change-Id: I5180c99c69ba394b8de895de411d0c5c4eccbbc4