Add logging for all inputs tests with version info.
Version.ts contains information about the current version of the virtual
keyboard. We should log that info for each test to simplify debugging
since we are dealing with 3 repos and versions are hard.

Example log:
[200~2022-09-12T02:34:23.612903Z [02:34:20.940] Input tools
version.txt: export const version = 'chromeos-input-tools_v_377_RC0';

Not the most elegant, but I think it is better to avoid trying to parse
the version.ts file.

BUG=None
TEST=CQ,DUT

Change-Id: Iee4acdf9dbe035ea26c6a237cc2d009e6f0e24d5