rmad: Add rmad to baseline
rmad forks out a root-level daemon to mount external disks. Design doc
with security approval: go/shimless-mount-executor.

We should add this to security.SandboxedServices so the test doesn't
fail. We plan to switch to a dedicated non-root user for the executor in
the future, and it's tracked in b:243855405.

BUG=b:246220000
TEST=tast -verbose <IP> security.*
     (There are some failures but nothing related to rmad)

Cq-Depend: chromium:3890472
Change-Id: I06109e91caa64ef87e31b6ffc7365e29511dbce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3892070
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>