platform.BootPerf: fix context timeout in preboot RPC call
This change fixes a bug in using the short context that is used to limit
the waiting time of CPU cool down to make the RPC call for shutting down
tlsdated. The bug makes the non-fatal error in CPU cool down timeout
fail the RPC call and the test.

BUG=b:242108590
TEST=tast run $IP_ADDR platform.BootPerf
TEST=./fast_build -T

Change-Id: Iae9061a8954d1874a319aab89164d4637ccc4054
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3890729
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>