wifi: P2PPerf test
Adding the test P2PPerf for testing the WiFi direct performance. The
test runs in the Nearbyshare pool where two DUTs exists. The main DUT is
configured as p2p group owner (GO) and the companion DUT is configured
as p2p client. After that, the p2p client is connected to the p2p group
owner (GO) network. Finally, the connection is tested by running TCP
Iperf where the p2p group owner (GO) is the Iperf server and the p2p
client is the Iperf client.

BUG=b:246387944
TEST=tast run -companiondut=cd1:localhost:2220 -var=router=localhost:2210 -var=pcap=localhost:2212 localhost:2200  wifi.P2PPerf
TEST=tast unit tests passed.

Change-Id: I7e88bd2936ae60baca19104d33bd7af6cbca191d