More robustly screenshare Google Docs in MeetCUJ presentation variants
Use the accessibility tree to find the exact button to click on to
screenshare the Google Docs tab for MeetCUJ tests that present. This
change addresses the incorrect assumption that the Google Docs tab is
always the third tab in the list of tabs to share the within the
"Share a Chrome tab" popup.

BUG=b:242578818
TEST=tast run ui.MeetCUJ.4p_present_notes_split
TEST=tast run ui.MeetCUJ.lacros_4p_present_notes_split

Change-Id: Idf5f1446acf67b75ca83d06551d8865305a417e9