network: virtualnet certs are installed separately from HTTPS server
Virtualnet package installs certificates when starting HTTPS server.
This is fine when only running one HTTPS server, but is not working
when more servers are needed. This patch makes it so that the
certificates are installed independently of the HTTPS server. This
also means that prior to starting the test's HTTPS server, the
certificates must be installed.

BUG=b:243620248
TEST=tast run <DUT> network.ShillCaptivePortalHTTP.*

Change-Id: Iba222bae1076c9fa1d0141e8c03d7f8f258922c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3891517
Reviewed-by: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>