network: virtualnet dnsmasq as DNS server
The current dnsmasq always runs as a DHCP server (and optionally
as a DNS server). This patch updates the behavior such that it is
possible for dnsmasq to only be a DNS server.

This patch also allows a virtualnet router to set a custom DNS
server addresses.

BUG=b:243620248
TEST=With the follow-up patches;
     tast run <DUT> network.DNSProxy.*;
     tast run <DUT> network.DNSProxyOverVPN.*;

Change-Id: I81e3bd5ffb7ffb517f7ea65d39f912383bbfe080
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3891518
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>