network: DNS proxy over VPN tests use virtualnet
The test is updated such that the primary network is using virtualnet.
Alongside that, the DNS and DoH server is updated to be a local server
also created through virtualnet.

The environment is set up as follows:
-------------    ----------    --------------
| interface | -> | router | -> | DNS server |
-------------    ----------    --------------
                     |         --------------------
	             --------> | VPN + DNS server |
                               --------------------

Alongside that, the block rules for VPN is also updated from FORWARD
to INPUT. This is done because the DNS server now lives inside the
VPN server. Previously the DNS server is outside the VPN server and
as such the DNS queries are forwarded by the VPN.

BUG=b:243620248
TEST=tast run <DUT> network.DNSProxyOverVPN.*

Change-Id: I885e249f81aac47dbf7a863853baba49531b613d