power: Make readFirstLine more robust
There is test flake in ui.TabSwitchCUJ evidently caused by readFirstLine
returning the empty string even though the file has content to be read.
This problem is hard to investigate because it happens unpredictably and
you can run the test a few times without ever seeing it.

I suspect that the problem is related to ignoring the error returned by
bufio.Reader.ReadString. There is a code comment explaining that the
error is intentionally ignored because "Reader.ReadString returns error
iff line does not end in \n" but I think that is a misunderstanding.
Read the Go API documentation carefully: "If ReadString encounters an
error before finding a delimiter, it returns the data read before the
error and the error itself (often io.EOF). ReadString returns err != nil
if and only if the returned data does not end in delim." This implies
that there are other potential reasons for an error besides just because
the delimiter is not found.

readFirstLine will use bufio.Scanner instead of bufio.Reader. This might
not solve the problem, but it should provide a helpful error message.

BUG=b:243567060
TEST=tast run <DUT> ui.TabSwitchCUJ on asurada/hayato

Change-Id: Ic0af0cd7e747776947737a0a6c9f2e89034372ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3892319
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>