arc.Fsp: Install Text app if not installed
arc.Fsp and arc.Fsp.vm are experiencing flakiness in lab DUTs as Text
app seems to be missing in some of them. Although this issue should be
investigated separately, this patch works around the issue by installing
the app in the beginning of the test if it's not installed.

BUG=b:244301416
TEST=tast run <DUT> arc.Fsp*

Change-Id: I693a65830da41c9206f01b566ad77fd3b7d54048