cellular_helper: Use context to enforce timeouts
I plan to add a function that allows us to pass in extra parameters
that we want a service to match in FindServiceForDevice. In order
to avoid combinatorial explosion of different variants of this
function, we can just get rid of the WithTimeout variant and ensure
that the user passes in a context with the timeout they want using
context.WithTimeout.

BUG=b:246395692
TEST=cellular.ShillCellularEnableAndConnect

Change-Id: I6b7297ee391456247ae584e3786e4273612ca1fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3892730
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>