graphics: Create Chamelium config file on disk
IGT on Chamelium requires a config file to get the chamelium IP.
Create a file that has all the required chamelium config and set an env
var for IGT to find it.

In the lab, the chamelium attached to each DUT has a hostname of
"DUT-HOSTNAME-chameleon".

If the test is running locally, pass a graphics.chameleonhostname
runtime var to `tast run`. If it's in the lab, it should get the DUT
hostname and append the suffix described above.

BUG=b:242878189
TEST=tast run -var=graphics.chameleonhostname=192:9992 localhost:22022  graphics.IgtChamelium.kms_chamelium

Change-Id: I3f8e2fb9d79f0a24aaadcf290c1f1d7ea0c2159f