network: Add timeout for ping expectation in some Routing* tests
Without the timeout, the test will only try ping only once for each
address. Although in this case we don't expect any change on
connectivity for the given network, but the routing itself can still be
affected by other networks (e.g., physical network is up/down), and that
will cause a transient connectivity issue.

This patch adds a timeout in the tests to avoid the flakiness caused by
that.

BUG=b:239753191
TEST=../platform/tast/fast_build.sh -T
TEST=tast run $DUT network.Routing*

Change-Id: I5e3bbc79c212c0389d8cb83db01cdf601e96b167
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3898157
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>