tast-tests: Fix policy.LacrosMainProfileLogin
The test was failing due to an unexpected new tab page opening after the
Settings page. Because of this the new tab page was an active tab and it
wasn't possible to see the content of the Settings page. Fix this by
waiting for the new tab to open before opening a Settings tab.

BUG=b:246306368
TEST=tast run <ip> accountmanager.LacrosMainProfileLogin*

Change-Id: Ib5a190c5bda02f0b50d7634ceb504d71b13d944f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3893868
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Georg Neis <neis@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>