Remove Speedometer from TaskSwitchCUJ
Speedometer caused ARC and Lacros to frequently crash on lower-end
devices. In order to get consistent data from ui.TaskSwitchCUJ*, we
removed it from this test, and if we want actual Speedometer data, that
could be done in a separate test.

BUG=b:243845235
TEST=Ran tast run <DUT> ui.TaskSwitchCUJ* 3 times, and verified that
each variant passed.

Change-Id: Ic7646a97fb49eb59d0c08a9655ad410cff6ecf90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3901111
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>