tast-tests: Get rid of polling from browserwatcher
Currently browser watcher is implemented in a way that starts
a new dedicated go routine and polling the procfs.

However, looking at its use, it is not necessary to poll.
Instead, we can just check the procfs when queried.
This will help us to improve the performance of the test. Specifically,
on measuring performance this will reduce the overhead of the Tast test
itself.
Experiemental power_Idle test porting showed us that this part caused
~20% increase of CPU usage of the CUJ, that is not negligible.
This fix tries to address it.

BUG=chromium:1364426
TEST=Ran login.Chrome

Change-Id: I71f2999ea12d7eaa768124d368b1729062868d33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3901038
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>