mojo: Add service manager fixture
Add fixture to track mojo_service_manager status. This can be used in
the tests require mojo. It will check if mojo_service_manager has
crashed during test or not.
This should be a parent fixture of ash-chrome. However, due to
b/235922792 the pid of service manager will change each time chrome
restart. This fixture can be added to those test with chrome after the
issue is fixed.

BUG=b:246912461
TEST=CQ

Change-Id: I1d7bff4500a1bd4f57f8658a4e2a675d90444704