faft: Improve on ECWakeOnCharge
Extend ec uart timeout so that there will be
enough time to run the ec hibernate command,
and receive output. Also, only check for the
tablet mode angles if DUT is a convertible.
Lastly, if lid is closed, skip long pressing
on pwr, and just wait for DUT to reach S3/G3.

BUG=b:192394835
TEST=manual on leased aleena, krane, scarlet, burnet

Change-Id: Ia5990323902e5e1497de621c4ebb7a03473349ee