tast-test: Enable system-proxy tests
The system-proxy tast tests require a managed proxy server configuration
pointing to a real web server. This used to be configured via the
ProxySettings policy, however this is not working anymore on Chrome OS.
Since the managed proxy server couldn’t be set, the test failed.

This CL updates the tast tests to use the deprecated proxy policies
ProxyServer and ProxyMode.

Additionally, this CL replaces the deprecated `ClassName` method with
`HasClass`.

BUG=b:241943012, b:243451525
TEST=Tests network.SystemProxyForArc and
     network.SystemProxyForSystemServices passed on DUT

Change-Id: Icb6afeceb642af549a1f3b1ac18248e77c4008b5