tast-tests: support multiple batteries in the Perfetto test
The battery data source now optionally add name to the counter name to
distinguish multiple batteries. For example, on a device with BAT0 and
hid-0001, the counter name will be 'batt.BAT0.capacity_pct' for BAT0's
capacity in percentage.

platform.PerfettoBatteryDataSource is updated to use the new (v29.0)
trace_processor_shell that handles battery counters of multiple
batteries.

BUG=b:232379863
TEST=`fast_build.sh -T`.
TEST=`tast run $IP_ADDR platform.PerfettoBatteryDataSource` on homestar

Change-Id: Idaf88afaecfc20be17e48d5ee363dca1cce11907