sched.CoreSchedTag: Add vulnerability dependency
We're still imprecisely targeting systems for this test. For one, some
Intel SMT systems (like Hatch/CometLake) appear to have core scheduling
enabled, but in fact they are not vulnerable and should not be running
this test.

Leverage the new 'CPUNeedsCoreScheduling' condition, which covers the
CPU vulnerabilities in question.

For some of these systems, they were previously passing this test
because of a quirk in our older kernels, in which we allowed the
scheduling tags/cookies to be set even when core scheduling was
disabled.

BUG=b:232380492
TEST=run it (eve=pass; scarlet=skip; hatch=skip)

Cq-Depend: chromium:3902630
Change-Id: Ifdc4698493a5c2880cfe0d797d9ca7d7fea69953
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3902941
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matt Davis <mattedavis@google.com>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>