[CrOS Cellular] Add tast tests to verify the roaming sub-label
Roaming toggle is a setting in network settings that can be used to
enable/disable roaming. Switching this setting on will display a
sub-label that is either 'Not currently roaming' or 'Currently roaming'
(depending on the type of the SIM).

These tests verifies the sub-label.

BUG=b:242066194
TEST=tast run localhost:2233 cellular.RoamingLabelOnNonRoamingSIM
     tast run localhost:2233 cellular.RoamingLabelOnRoamingSIM

Change-Id: I31fd6f9658a6fe59845e8b8dbf699330681c1d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3903165
Tested-by: Nikhil Nayunigari <nikhilcn@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Nikhil Nayunigari <nikhilcn@google.com>