tast-tests: Validate auto install of blocked app for unicorn
Currently blocking of app install is enforced by the Play Store and it
depends on user settings being downloaded in the background.

If the install is atttempted before user settings are downloaded the
install button is enabled and the download attempt succeeds but ARC DPC
uninstalls the app soon after installation as it is in the list of
blocked apps.

The test was failing only on presence of enabled install button but now
it is going to attempt the install and validate that the app is
immediately uninstalled. If the app is not uninstalled then the test
will fail.

BUG=b:237023275
TEST=tast run DUT arc.UnicornDisabledApps

Change-Id: I3bfb47547a2dc5c104378a9cff74d1e139571c02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3903166
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>