video/platform_decoding: Enable HEVC decode tests for SC7180 devices
This CL changes the dependency for HEVC V4l2 Stateful decoding tests to
use the DUT model. Eligible models use the trogdor or strongbad
platform.

This is a stopgap measure to enable testing until the Chrome HEVC
decoder can be finished and enabled for test builds. After which, the
autotest video capability can be enabled for the devices and this commit
can be reverted.

BUG=b:227248846
BUG=b:232255167
TEST=video.PlatformDecoding.v4l2_hevc_main test runs and passes on
TEST=lazor and coachz devices

Change-Id: I7bbba485c75af36723a77c0aff965ddcfb50266e