video/platform_decoding: Enable HEVC decode tests for SC7180 devices
This CL changes the dependency for HEVC V4l2 Stateful decoding tests to
use the DUT platform. Eligible platforms are trogdor and strongbad. This
is a stopgap measure to enable testing until the Chrome HEVC decoder can
be finished and enabled for test builds. After which, the autotest
capability can be enabled.

BUG=b:227248846
BUG=b:232255167
TEST=video.PlatformDecoding.v4l2_hevc_main test runs and passes on
TEST=trogdor and strongbad devices

Change-Id: I7bbba485c75af36723a77c0aff965ddcfb50266e