arc: Create perfetto package
This CL move perfetto.go to perfetto dir and change to use
perfetto as package name.

Perfetto is a independent tool set, and there will be some follow
up functions for other perfetto tools and commands. So make it as
a new package to prevent namespace confusing.

BUG=b:245042021
TEST=build tast test

Change-Id: I2ae1890cb8914ddfa1efbbf49c10b5d8df300458
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3906506
Commit-Queue: Shengsong Tan <sstan@chromium.org>
Tested-by: Shengsong Tan <sstan@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>