assistant: Add test to check shutdown crash
There was a bug that libassistant DLC crash during shutdown on device.
This patch adds a tast test to detect possible shutdown crash.

The method in this patch is simply to checkout `chrome.PREVIOUS` log, if
there is a line containing `chrome_crashpad_handler`, then there may be
a crash. We could possibly check other dump file as well in the future.

BUG=b:247918343
TEST=This new tast test

Change-Id: I8c5bf43a1747ffe7ef46cac7588f8e472f4c7bbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3911322
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Yuki Awano <yawano@google.com>
Tested-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Tao Wu <wutao@chromium.org>