faft: Improve on fixture.go
Earlier we found some DUTs stuck at G3 during reboot
to a specified mode. Check for the DUT's power state
and document it in the error message. This information
would help us in filtering Stainless results. Also,
Increase the PreTest, PostTest, and TearDown timeouts
to account for DUTs that might take longer time to
finish a power cycle.

BUG=b:200305315
TEST=ran a few faft tests on Juniper, Storo360, and Eldrid

Change-Id: I03834aa478f3d84fafca24163ce7fa80a73c1b07