tast-tests: Add lacros.ShelfLaunchOmaha
This is a rewrite of lacros.ShelfLaunch.omaha to support the version
skew smoke testing against the valid version skews served in Omaha.

This test runs multiple subtests to check all supported stateful-lacros
channel depending on the current OS channel it runs on.
For example, if it kicks off on the OS canary (before a new branch is
cut for a milestone) it will find the canary lacros channel. If on the
OS stable, it will do all the (dev|beta|stable) lacros channels.

Note that this test may fail for a certain period of time on the OS dev
when there is no newer stateful-lacros served in Omaha than rootfs-lacros.
This is intended until a new stateful-lacros is rolled out after the
manual qualification.

BUG=chromium:1367120, chromium:1316191, b:246818834
TEST=tast run $DUT lacros.ShelfLaunchOmaha (on the OS canary, dev, beta, stable)

Change-Id: I1da4d4c9b9f0ee7ff327bc54de50c55f16aab606