ui: Begin WindowArrangementCUJ with left tab active
WindowArrangementCUJ will begin with the left tab active for 2 reasons:
1. So that when the workflow loops, it is consistent with the beginning
of the whole test. This matters because trace data is collected for the
first minute and assumed to be representative of the whole test.
2. Because it is a better performance measurement with a video showing
in the window during the move, resize, and state changes.

BUG=b:247634722
TEST=tast run <DUT> ui.WindowArrangementCUJ.*

Change-Id: Ie893c166148c2fdd43f8c3b437b10433a80c5406
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3907913
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>