ui: Use 90% zoom or less, in OverviewDragWindowPerf
The tests require that browser windows can be snapped in split view.
This requires that the work area length is at least twice the minimum
length of a browser window. To ensure this, the tests will ensure that
the display zoom factor is 90% or lower.

BUG=chromium:1364955
TEST=tast run <DUT> ui.OverviewDragWindowPerf.*

Change-Id: I01c05d7f1ed96a9ce8209152d5dd20675ee05dc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3908277
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: David Welling <davidwelling@google.com>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Navin Ramsaroop <ramsaroop@google.com>