Fix and simplify AppInfoFiles tast test
Fix and simplify a broken tast test. Previously, the test checked the
contents of the Files app, but since Files app is a system web app and
therefore doesn't have an App Info page in App Management, which caused
test failure. We've changed the test to check for the App Info page on
the Web Store app (which is pre-installed). The test originally opens
up the App Info page from both the shelf and the launcher, but this
change only does the check from the launcher since the page is the
same from either entry point.

BUG=chromium:1359166
TEST=tast run apps.AppInfoWebStore

Change-Id: Id59d339aa5241cb40a1431d66e9d2fd88843d958
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3909072
Commit-Queue: Vivian Pao <vpao@google.com>
Tested-by: Vivian Pao <vpao@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>