media: Stop chrome from executing fwupd
Chrome browser executes fwupd when chrome browser restarts (e.g. a
user logs in). The fwupd restarts chromeos service daemons. powerd is
one of the restarted daemons. This is problematic because we stop
powerd in the beginning of the performance tests but it restarts in
logging in.
This CL avoids this problem by stopping chrome from executing fwupd
by --disable-features=FirmwareUpdaterApp.

BUG=b:246431881
TEST=powerd doesn't restart by fwupd in video.PlaybackPerf.h264_1080p_30fps_hw_oopvd

Change-Id: I9af7dc44ca73a4028092a0b75c733993e3f5ce80