Tast: Allow ARC to get login info in demo mode test
chrome.ARCEnabled() would emit Chrome arg
"--disable-arc-opt-in-verification", which would cause ash
arc_auth_service to go a differnt code path when fetching primary
account into. This eventually results in Play Store not having primary
account login info, thus would ask user to log in.

This CL switches to a different option ARCSupported(), which does not
emit this flag.

This change is prerequisite of a demo mode test for ARC.

BUG=b:235602920
TEST=Play Store is usable after test. tast run DUT demomode.SetUp

Change-Id: If78af258a8a91f4c3d2dd9b06133c9bd6f210bf8