tast-tests: Stabilize launcher.HideContinueSection*
Remove the chrome flag LauncherHideContinueSection, which has been
on-by-default since M-105 and isn't needed any more.

Instead of opening local files with the Files app to populate the
continue section, pass ForceShowContinueSection to force chrome to
populate it. This is both more reliable and faster (decreases runtime of
HideContinueSectionClamshell by 7 seconds per run).

In HideContinueSectionClamshell, dismiss the sorting nudge then close
and reopen the launcher before checking for the continue section. This
forces a fetch to the search results provider.

Both HideContinueSectionTablet and HideContinueSectionClamshell now pass
10/10 times locally on eve.

BUG=chromium:1364765
TEST=tast run <dut> launcher.HideContinueSection*

Change-Id: If38278ea425bf2d8c8695e7268e40587a696978a