Fix OS variations smoke test
The test started failing for 2 reasons. First, injecting the test seed
became flaky. It seems we can't set the variations prefs in the first
few moments immediately after Chrome starts. We can fix this by
repeating it at an interval until it sticks. Second, the pixel test was
failing due to a tool-tip covering up part of the browser window. This
is fixed by moving the mouse away from the browser before starting the
screen diff.

BUG=b:243451722
TEST=tast run <dut> ui.ChromeVariationsSmoke

Change-Id: I80412f2647d693fa83c3e8805e1388fb7cd3426c