runtimeprobe: Add verify sandbox test
Add test to verify the sandbox of runtime probe by comparing the result
from runtime probe and factory runtime probe. The result should be the
same, but the factory runtime probe don't run in sandbox.
Split the functions into groups so some of them could be promote first.
We expect to collect some failure for those unstable function to fix the
sandbox issue before promote.

BUG=b:236666921
TEST=tast run $DUT runtimeprobe.*

Change-Id: Id3ff0676a346574474c2f444d28f88ea848b84a0