Move Demo Mode setup test behavior to fixture for re-use
Since the demomode.SetUp test would now bw empty, this CL
removes it, and instead adds a test to assert that the core
Demo Chrome Apps have launched, as a proof-of-concept
for using the fixture for further testing within the
launched Demo Session.

The one new thing added to the SetUp functionality
is overriding the DeveloperToolsEnabled policy,
which is disabled in prod demo mode, but needs
to be enabled for Tast so that we can connect
to the test API extension.

BUG=b:248388861
TEST=tast -verbose run <device-ip> demomode.ChromeApps

Change-Id: I60c24d29acb57c7fa88b7f058f1f71353cd355d6