Projector: Make the CreationFlow Tast test less stringent
This CL does the following:
1) Remove the check for pre-existing screencast items before starting
the test. Given the number of concurrently running tests generating
screencasts, it is unreasonable for this test to expect no pre-
existing screencast items before starting.
2) Remove dismissing the tutorial videos. Since there can be any
number of pre-existing screencasts on the gallery, and the tutorial
videos disappear after three or more screencasts, we should no longer
dismiss the tutorial video panel. This step was also unnecessary for
the test to pass.

BUG=b:248298346
TEST=tast run -buildbundle=cros localhost:2222 projector.CreationFlow

Change-Id: I7a4124d0a93d5b3c08ff7b1a4617d8825a720c85