mtbf: Fix arc.AudioRecordAndPlayBack
The app will not achieve idle state while it's voice recording, thus, the app is not able to interact through the uiautomator during that time which makes the test unreliable.

This CL fixes the issue by applying the usage of uidetection to detect UI status.
Also, this CL refines the logic of setting up the output directory of a recording file to avoid test influences by the ads.

BUG=b:247435763
TEST=tast run <DUT> arc.AudioRecordAndPlayback

Change-Id: I59e7ccbdf419b79c25a208e396c1a62f2ccfdc13