tast: Fix order dependent failure in inputs.VirtualKeyboardEnglishSettings
VirtualKeyboardEnglishSettings fails when running after
VirtualKeyboardAutocorrect.en_us_tablet. This is because the latter
disables autocap, but the former relies on autocap.

Always run VirtualKeyboardEnglishSettings with a fresh Chrome session so
that autocap tests will work.

Also, there's an order dependency within the subtests of
VirtualKeyboardEnglishSettings: it assumes that autocap will always be
enabled by default, so it only disables autocap, never enables it. Fix
this by always setting the autocap setting, regardless of whether it is
enabled or disabled.

BUG=b:248455858
TEST=tast run -checktestdeps=false localhost:2222 inputs.VirtualKeyboardAutocorrect.en_us_tablet inputs.VirtualKeyboardEnglishSettings

Change-Id: Ic6d324655cb36551a317a787539298b493c9e6fc