[QuickStart] Add crossdeviceNoSignIn fixture for QuickStart e2e tests
This CL adds a fixture that will be associated with an RF box and an
Android Device that can be controlled by tast. Thix fixture is
different than existing fixtures because it doesn't sign in to GAIA
during fixture setup.

go/cros-quick-start-e2e-test-plan

BUG=b:234655072
TEST=manually tested quickstart.SingleAccountOnboarding that brings
up this fixture.

Change-Id: I20485c518e0f84f16d6b75066e48e1955dec8bd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3916585
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Commit-Queue: Jason Rhee <jasonrhee@google.com>
Tested-by: Jason Rhee <jasonrhee@google.com>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>