ui: Update TabSwitchCUJ to incorporate mouse input
After each tab switch, pause and scroll down and up using the keyboard
and mouse scroll wheel. Additionally, use the mouse to move to each tab
and activate it, mimicking an actual user experience.

In completing the above changes, a few improvements were made. Some of
the variable names were standardized to how most of our other CUJ tests
are written for help with readability and future development. The test
is also more accurate in its 10 minute duration cap -- it checks after
each individual tab switch if 10 minutes has passed. Lastly, the test
now properly removes the "New Tab" window in the Lacros variant, so the
Ash and Lacros test variants are equivalent.

A shortened version of the updated TabSwitchCUJ is found here, with the
core part of the test starting at 2:15:
https://drive.google.com/file/d/1zOcv4-RM3Bx3iVNv3--lR0tbKa4VmupS/view?usp=sharing&resourcekey=0-6BpTAT-cSuvfEcFNtyyZEQ

BUG=b:241841511
TEST=tast run <DUT> ui.TabSwitchCUJ*

Change-Id: Ic6698972c17b20b9654340e7bcf31a3b8ada6096
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3916589
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Tested-by: Navin Ramsaroop <ramsaroop@google.com>