Projector: Only delete old screencast items
Currently, DeleteScreencastItems() deletes all the screencasts on the
gallery, interfering with concurrently running tests by deleting their
screencasts. This CL modifies DeleteScreencastItems() to only delete
screencast items after the 20th one. This deletion mechanism is simpler
than trying to parse the created time and calculating the age of the
screencast. It's still important to delete screencast items to avoid
taking up Drive quota, but we should leave a buffer of recently
created screencasts to avoid interfering with other tests.

BUG=b:248298346
TEST=tast run -buildbundle=cros localhost:2222 projector.*

Change-Id: I0a1b883f922716b08d38a3aef8aa8c41e6f8f4b7