tast: CrasRecordQuality: Remove RMS threshold checking for CheckRecordingQuality
We used to test if the RMS of the recording is too large to detect
intrinsic static noise. However, this cannot distinguish between
intrinisc or external noise, which make this this test flaky depending
on the environment inside the lab.

Removing the threshold checking and only ensuring the recording is not
zero can help to improve the stability of this test.

BUG=b:230695223
TEST=tast run <dut> CrasRecordQuality

Change-Id: I4443898125e850ce66060904c75d5ff057bf006a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3915034
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>